Added missing camelCase option for locals loader
Fixes https://github.com/webpack/css-loader/pull/149#issuecomment-150114155
