+++ lib/loader.js
@@ -4,9 +4,9 @@
 */
 var path = require("path");
 var loaderUtils = require("loader-utils");
-var camelCase = require("lodash.camelcase");
 var processCss = require("./processCss");
 var getImportPrefix = require("./getImportPrefix");
+var compileExports = require("./compile-exports");
 
 
 module.exports = function(content, map) {
@@ -82,21 +82,11 @@ module.exports = function(content, map) {
 			return "\" + require(" + loaderUtils.stringifyRequest(this, urlRequest) + ") + \"";
 		}.bind(this));
 
-		var exportJs = "";
-		if(Object.keys(result.exports).length > 0) {
-			var boundImportItemMatcher = importItemMatcher.bind(this);
-			exportJs = Object.keys(result.exports).reduce(function(res, key) {
-				var valueAsString = JSON.stringify(result.exports[key]);
-				valueAsString = valueAsString.replace(result.importItemRegExpG, boundImportItemMatcher);
-				res.push("\t" + JSON.stringify(key) + ": " + valueAsString);
-				if (camelCaseKeys) {
-					res.push("\t" + JSON.stringify(camelCase(key)) + ": " + valueAsString);
-				}
-				return res;
-			}, []).join(",\n");
-			exportJs = "exports.locals = {\n" + exportJs + "\n};";
-		}
 
+		var exportJs = compileExports(result, importItemMatcher.bind(this), camelCaseKeys);
+		if (exportJs) {
+			exportJs = "exports.locals = " + exportJs;
+		}
 
 		var moduleJs;
 		if(query.sourceMap && result.map) {
