+++ lib/localsLoader.js
@@ -5,6 +5,7 @@
 var loaderUtils = require("loader-utils");
 var processCss = require("./processCss");
 var getImportPrefix = require("./getImportPrefix");
+var compileExports = require("./compile-exports");
 
 
 module.exports = function(content) {
@@ -12,6 +13,7 @@ module.exports = function(content) {
 	var callback = this.async();
 	var query = loaderUtils.parseQuery(this.query);
 	var moduleMode = query.modules || query.module;
+	var camelCaseKeys = query.camelCase || query.camelcase;
 
 	processCss(content, null, {
 		mode: moduleMode ? "local" : "global",
@@ -33,15 +35,11 @@ module.exports = function(content) {
 				"[" + JSON.stringify(importItem.export) + "] + \"";
 		}
 
-		var exportJs = "";
-		if(Object.keys(result.exports).length > 0) {
-			exportJs = Object.keys(result.exports).map(function(key) {
-				var valueAsString = JSON.stringify(result.exports[key]);
-				valueAsString = valueAsString.replace(result.importItemRegExpG, importItemMatcher.bind(this));
-				return "\t" + JSON.stringify(key) + ": " + valueAsString;
-			}.bind(this)).join(",\n");
-			exportJs = "module.exports = {\n" + exportJs + "\n};";
+		var exportJs = compileExports(result, importItemMatcher.bind(this), camelCaseKeys);
+		if (exportJs) {
+			exportJs = "module.exports = " + exportJs;
 		}
+
 
 		callback(null, exportJs);
 	}.bind(this));
